services/users/requirements.txt
coverage==4.5.4
flake8===3.7.8
Flask==1.1.1
flask-bcrypt==0.7.1
flask-cors==3.0.8
flask-debugtoolbar==0.10.1
flask-migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
gunicorn==19.9.0
pyjwt==1.7.1
psycopg2-binary==2.8.3
